fam00200, FAT00100, FAt00700, FAT01100

+++ BIMASAKTI_11/1.00/PROGRAM/BS Program/SOURCE/COMMON/FA/FAT00100COMMON/DTOs/FAT00100GetTransAssetListResultDTO.cs
@@ -21,7 +21,7 @@ namespace FAT00100Common.DTOs
         public string CLOCAL_CURRENCY_CODE { get; set; } = string.Empty;
         public string CBASE_CURRENCY_CODE { get; set; } = string.Empty;
         public decimal NINIT_COST { get; set; }
-        public short IQTY { get; set; }
+        public int IQTY { get; set; }
         public string CUNIT { get; set; } = string.Empty;
         public decimal NLINIT_COST { get; set; }
         public decimal NBINIT_COST { get; set; }
